# docs/custom-gpt-actions/openapi.yaml
openapi: 3.1.0
info:
  title: Life OS Custom GPT Actions
  version: 1.0.0
servers:
  - url: https://YOUR-VERCEL-DOMAIN.vercel.app
paths:
  /api/actions/context:
    get:
      operationId: getLifeOSContext
      summary: Get the current user's Life OS goals and aliases
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user's context
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      actionCredential:
                        type: string
                  goals:
                    type: array
                    items:
                      type: object
                  aliases:
                    type: array
                    items:
                      type: object
        "401":
          description: Invalid or missing action credential
        "429":
          description: Rate limit exceeded
  /api/actions/life-event:
    post:
      operationId: recordLifeEvent
      summary: Record one structured Life OS event for the authenticated credential owner
      description: Do not include userId in the request body. The backend derives user ownership from the Bearer token.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - type
                - rawText
                - confidence
              properties:
                type:
                  type: string
                  enum: [activity, task, goal, reminder, inbox]
                rawText:
                  type: string
                  maxLength: 500
                confidence:
                  type: number
                  minimum: 0
                  maximum: 1
                goal:
                  type: object
                  additionalProperties: false
                  properties:
                    title:
                      type: string
                      maxLength: 500
                    category:
                      type: string
                      maxLength: 500
                    parentTitle:
                      type: string
                      maxLength: 500
                    metricType:
                      type: string
                      enum: [duration, count, milestone]
                    aliases:
                      type: array
                      maxItems: 12
                      items:
                        type: string
                        maxLength: 500
                summary:
                  type: string
                  maxLength: 500
                metric:
                  type: object
                  additionalProperties: false
                  properties:
                    type:
                      type: string
                      enum: [duration, count, milestone]
                    value:
                      type: number
                      exclusiveMinimum: 0
                      maximum: 100000
                    unit:
                      type: string
                      enum: [minute, hour, count]
                date:
                  type: string
                  pattern: "^\\d{4}-\\d{2}-\\d{2}$"
                task:
                  type: object
                  additionalProperties: false
                  properties:
                    title:
                      type: string
                      maxLength: 500
                    dueAt:
                      type: string
                      format: date-time
                    priority:
                      type: string
                      enum: [low, normal, high]
                reminder:
                  type: object
                  additionalProperties: false
                  properties:
                    remindAt:
                      type: string
                      format: date-time
                    repeatRule:
                      type: string
                      enum: [none, daily, weekly]
                suggestedTypes:
                  type: array
                  maxItems: 5
                  items:
                    type: string
                    enum: [task, activity, goal, reminder, inbox]
                reason:
                  type: string
                  maxLength: 500
      responses:
        "200":
          description: Life event recorded
        "400":
          description: Invalid payload
        "401":
          description: Invalid or missing action credential
        "429":
          description: Rate limit exceeded
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer